Starting Workout Unit Tests

+++ lib/home/workouts/tags.ts
@@ -1,13 +1,14 @@
 "use server";
 import prisma from "@/client";
 import { z } from "zod";
+import { uuidSchema } from "@/lib/global/zod";
+import { Workout } from "@/lib/home/workouts/workouts";
 import {
    sendSuccessMessage,
    sendErrorMessage,
    sendFailureMessage,
    VitalityResponse
 } from "@/lib/global/response";
-import { uuidSchema } from "@/lib/global/zod";
 
 export type Tag = {
   user_id: string;
@@ -34,24 +35,11 @@ const workoutTagSchema = z.object({
 
 export async function fetchWorkoutTags(userId: string): Promise<Tag[]> {
    try {
-      const tags = await prisma.workout_tags.findMany({
+      return await prisma.workout_tags.findMany({
          where: {
             user_id: userId
          }
       });
-
-      const result: Tag[] = [];
-
-      for (let i = 0; i < tags.length; i++) {
-         result.push({
-            user_id: tags[i].user_id,
-            id: tags[i].id,
-            title: tags[i].title,
-            color: tags[i].color
-         });
-      }
-
-      return result;
    } catch (error) {
       return [];
    }
@@ -61,72 +49,90 @@ export async function addWorkoutTag(tag: Tag): Promise<VitalityResponse<Tag>> {
    const fields = workoutTagSchema.safeParse(tag);
 
    if (!fields.success) {
-      // Return the field errors
-      const errors = fields.error.flatten();
-
-      // Only error should be proposed tag title not being a valid length
-      if (errors.fieldErrors.title) {
-         return sendErrorMessage("Invalid workout tag fields", {
-            title: errors.fieldErrors.title ?? [""]
-         });
-      }
+      return sendErrorMessage(
+         "Invalid workout tag fields",
+         fields.error.flatten().fieldErrors
+      );
    }
 
    try {
+      const existingTag = await prisma.workout_tags.findFirst({
+         where: {
+            title: tag.title.trim(),
+            user_id: tag.user_id.trim()
+         }
+      });
+
+      if (existingTag) {
+         return sendErrorMessage("Workout tag title already exists", {
+            title: ["Workout tag title already exists"]
+         });
+      }
+
       const newTag: Tag = await prisma.workout_tags.create({
          data: {
             user_id: tag.user_id,
-            title: tag.title,
-            color: tag.color
+            title: tag.title.trim(),
+            color: tag.color.trim()
          }
       });
 
       return sendSuccessMessage("Successfully added new workout tag", newTag);
    } catch (error) {
-      if (
-         error.code === "P2002" &&
-         error.meta?.modelName === "workout_tags" &&
-         error.meta?.target?.includes("user_id") &&
-         error.meta?.target?.includes("title")
-      ) {
-         // Workout tags must be unique
-         return sendErrorMessage("Workout tag already exists", {
-            search: ["Workout tag already exists"]
-         });
-      } else {
-         return sendFailureMessage(error);
-      }
+      return sendFailureMessage(error);
    }
 }
 
 export async function updateWorkoutTag(
    tag: Tag,
-   method: "update" | "delete",
+   method: "update" | "delete"
 ): Promise<VitalityResponse<Tag>> {
    const fields = workoutTagSchema.safeParse(tag);
-
-   // Handle invalid tag id's and user fields
-   if (tag.id === undefined) {
-      return sendErrorMessage("Missing Workout Tag ID", null);
+   // Handle missing tag-related id's or invalid user fields
+   if (tag.user_id.trim() === "") {
+      return sendErrorMessage("Missing user ID", null);
+   } else if (tag.id.trim() === "") {
+      return sendErrorMessage("Missing workout tag ID", null);
    } else if (!fields.success) {
       return sendErrorMessage(
          "Invalid workout tag fields",
-         fields.error.flatten().fieldErrors,
+         fields.error.flatten().fieldErrors
       );
    }
 
    try {
-      // Handle update/delete workout tags
+      // Ensure the workout tag exists
+      const existingTag = await prisma.workout_tags.findFirst({
+         where: {
+            id: tag.id.trim(),
+            user_id: tag.user_id.trim()
+         }
+      });
+
+      if (!existingTag) {
+         return sendErrorMessage(
+            "Workout tag does not exist based on user ID and/or tag ID",
+            null
+         );
+      }
+
+      // Update or remove the workout tag
       switch (method) {
          case "update":
             const newTag = await prisma.workout_tags.update({
                where: {
                   id: tag.id
                },
-               data: tag
+               data: {
+                  title: tag.title.trim(),
+                  color: tag.color.trim()
+               }
             });
 
-            return sendSuccessMessage("Successfully updated workout tag", newTag);
+            return sendSuccessMessage(
+               "Successfully updated workout tag",
+               newTag
+            );
          case "delete":
             const deletedTag = await prisma.workout_tags.delete({
                where: {
@@ -134,27 +140,40 @@ export async function updateWorkoutTag(
                }
             });
 
-            return sendSuccessMessage("Successfully deleted workout tag", deletedTag);
-         default:
-            return sendErrorMessage(
-               "Invalid Workout Tag Update Method",
-               null
+            return sendSuccessMessage(
+               "Successfully deleted workout tag",
+               deletedTag
             );
+         default:
+            return sendErrorMessage("Invalid Workout Tag Update Method", null);
       }
    } catch (error) {
-      if (
-         error.code === "P2002" &&
-         error.meta?.modelName === "workout_tags" &&
-         error.meta?.target?.includes("user_id") &&
-         error.meta?.target?.includes("title")
-      ) {
-         // Workout tags must be unique by their title
-         return sendErrorMessage("Workout tag title already exists", {
-            title: ["Workout tag title already exists"]
-         },
-         );
-      } else {
-         return sendFailureMessage(error);
-      }
+      return sendFailureMessage(error);
    }
+}
+
+export async function getAppliedWorkoutTagUpdates(
+   existingWorkout,
+   newWorkout: Workout
+): Promise<{ adding: string[]; removing: string[] }> {
+   // Extract existing applied tag IDs
+   const existing: Set<string> = new Set(
+      existingWorkout?.workout_applied_tags.map((tag) => tag.tag_id) || []
+   );
+
+   // Determine tags ID's to add and remove from existing workout
+   const adding: Set<string> = new Set(newWorkout.tagIds);
+
+   const addingTags: string[] = Array.from(adding).filter(
+      (id) => !existing.has(id)
+   );
+
+   const removingTags: string[] = Array.from(existing).filter(
+      (id) => !adding.has(id)
+   );
+
+   return {
+      adding: addingTags,
+      removing: removingTags
+   };
 }
